refactoring scripts
