+++ contracts/lockdrop/src/contract.rs
@@ -1097,9 +1097,9 @@ pub fn handle_claim_rewards_and_unlock_for_lockup(
                     "No staking rewards available to claim, lockdrop reward already claimed!",
                 ));
             }
+        } else if user_info.astro_transferred && !withdraw_lp_stake {
+            return Err(StdError::generic_err("No rewards available to claim!"));
         }
-    } else if user_info.astro_transferred && !withdraw_lp_stake {
-        return Err(StdError::generic_err("No rewards available to claim!"));
     }
 
     cosmos_msgs.push(
